add support for standard calls
